managed to print the words stored in a WordList.
notes:

there is a bit of an issue with using `fgets` for this. since
`fgets` reads from a stream with size BUFFER as the size, it does not
check for newline characters. ideally, we would like to check each
character in the stream for newlines so we can split it up according
to new lines.